Remove eventually module
